Refactor: rename helpers directory to widgets
