+++ src/components/error_message/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { withTranslation } from 'react-i18next';
 
-import IrmaAppBar from '../../helpers/irma_app_bar';
-import Column from '../../helpers/column';
-import IrmaButton from '../../helpers/irma_button';
+import IrmaAppBar from '../../widgets/irma_app_bar';
+import Column from '../../widgets/column';
+import IrmaButton from '../../widgets/irma_button';
 import { connect } from 'react-redux';
 
 const mapStateToProps = (state) => {
