+++ src/widgets/chevron-left-icon.js
