+++ src/widgets/chevron-right-icon.js
