+++ src/widgets/column.js
