+++ src/widgets/irma_app_bar.js
