+++ src/widgets/irma_app_bar.scss
