+++ src/widgets/irma_button.js
