+++ src/widgets/irma_button.scss
