+++ src/widgets/irma_modal.js
