+++ src/widgets/irma_modal.scss
