+++ src/widgets/irma_table.js
