+++ src/widgets/irma_table.scss
