+++ src/widgets/loading_spinner.js
