+++ src/widgets/loading_spinner.scss
