+++ src/widgets/logout_icon.js
