+++ src/widgets/success_icon.js
